move parsing code into its own class and module
parser.SlackParser
